\name{LASSO-package}
\alias{LASSO-package}
\alias{LASSO}
\docType{package}
\title{
\packageTitle{LASSO}
}
\description{
\packageDescription{LASSO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{LASSO}
\packageIndices{LASSO}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{LASSO}

Maintainer: \packageMaintainer{LASSO}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
