\name{Lasso-package}
\alias{Lasso-package}
\alias{Lasso}
\docType{package}
\title{
\packageTitle{Lasso}
}
\description{
\packageDescription{Lasso}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{Lasso}
\packageIndices{Lasso}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{Lasso}

Maintainer: \packageMaintainer{Lasso}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
